+++ example/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:expansion_tile_group_example/route_named.dart';
 import 'package:expansion_tile_group_example/use_cases/fantasy/fantasy_page.dart';
-import 'package:expansion_tile_group_example/use_cases/group/expand_only_with_listview.dart';
+import 'package:expansion_tile_group_example/use_cases/group/custom_group_with_controller.dart';
 import 'package:flutter/material.dart';
 
 import 'use_cases/use_cases.dart';
@@ -50,9 +50,9 @@ class _MyAppState extends State<MyApp> {
           case RouteNamed.expandAlwaysCurrent:
             return MaterialPageRoute(
                 builder: (context) => const ExpandAlwaysCurrentPage());
-            case RouteNamed.expandOnlyWithListView:
+            case RouteNamed.customGroupWithController:
             return MaterialPageRoute(
-                builder: (context) => const ExpandOnlyWithListView());
+                builder: (context) => const CustomGroupWithController());
           case RouteNamed.fantasyPage:
             return MaterialPageRoute(builder: (context) => const FantasyPage());
           default:
@@ -171,7 +171,7 @@ class HomePage extends StatelessWidget {
           _buildRow(
             context,
             title: 'Custom Group with controller',
-            routeName: RouteNamed.expandOnlyWithListView,
+            routeName: RouteNamed.customGroupWithController,
           ),
         ],
       ),