+++ example/lib/route_named.dart
@@ -4,7 +4,7 @@ class RouteNamed {
   static const String listenGroupItemChanged = '/listen_group_item_changed';
   static const String expandOnlyCurrent = '/expand_only_current';
   static const String expandAlwaysCurrent = '/expand_always_current';
-  static const String expandOnlyWithListView = '/expand_only_with_listview';
+  static const String customGroupWithController = '/custom_group_with_controller';
   static const String collapseAll = '/collapse_all';
   static const String expandAll = '/expand_all';
   static const String expandOrCollapseAll = '/expand_or_collapse_all';