+++ example/lib/use_cases/fantasy/example_action_card.dart
@@ -46,10 +46,8 @@ These buttons control the card above!""",
           style: Theme.of(context).textTheme.bodyMedium!.copyWith(fontSize: 16),
         ),
       ),
-      ButtonBar(
-        alignment: MainAxisAlignment.spaceAround,
-        buttonHeight: 52.0,
-        buttonMinWidth: 90.0,
+      Row(
+        mainAxisAlignment: MainAxisAlignment.spaceAround,
         children: <Widget>[
           TextButton(
             style: flatButtonStyle,