+++ example/lib/use_cases/group/listen_changed_item_page.dart
@@ -88,26 +88,26 @@ class _ListenChangedItemPageState extends State<ListenChangedItemPage> {
                 ),
               ],
             ),
-            ExpansionTileOutlined(
-              title: const Text('ExpansionTile 1'),
+            const ExpansionTileOutlined(
+              title: Text('ExpansionTile 1'),
               expendedBorderColor: Colors.red,
-              children: const [
+              children: [
                 Text(
                     '''Duis rutrum convallis erat et ultrices. Morbi a luctus ligula, at varius ligula. Nam mollis sapien ac nunc hendrerit consequat. Cras posuere metus felis, at pellentesque sem ornare id. Praesent ut nunc aliquam, dictum felis eu, congue metus. Nunc vitae elit eros. In eu dui pharetra, varius metus a, efficitur eros.'''),
               ],
             ),
-            ExpansionTileOutlined(
-              title: const Text('ExpansionTile 2'),
+            const ExpansionTileOutlined(
+              title: Text('ExpansionTile 2'),
               expendedBorderColor: Colors.green,
-              children: const [
+              children: [
                 Text(
                     '''Duis rutrum convallis erat et ultrices. Morbi a luctus ligula, at varius ligula. Nam mollis sapien ac nunc hendrerit consequat. Cras posuere metus felis, at pellentesque sem ornare id. Praesent ut nunc aliquam, dictum felis eu, congue metus. Nunc vitae elit eros. In eu dui pharetra, varius metus a, efficitur eros.'''),
               ],
             ),
-            ExpansionTileOutlined(
-              title: const Text('ExpansionTile 3'),
+            const ExpansionTileOutlined(
+              title: Text('ExpansionTile 3'),
               expendedBorderColor: Colors.blue,
-              children: const [
+              children: [
                 Text(
                     '''Duis rutrum convallis erat et ultrices. Morbi a luctus ligula, at varius ligula. Nam mollis sapien ac nunc hendrerit consequat. Cras posuere metus felis, at pellentesque sem ornare id. Praesent ut nunc aliquam, dictum felis eu, congue metus. Nunc vitae elit eros. In eu dui pharetra, varius metus a, efficitur eros.'''),
               ],